Karabiner: Mapping HyperKey(F14) to Command.
The F14 key is set on my FreeStyle Pro keyboard to the very convenient
to press physical Left side Space key.  This key was being used as a
Hyper key with the mapping being done using the Hammerspoon application.

However I have found it is better to use this key as the Command key,
when using MacOS, hence updating Karabiner to perform this mapping.