Better dual purpose Capslock key.
Instead of using Hammerspoon, doing it all via Karabiner-elements.  The
Hammerspoon implementation was too slow, and had corner cases where
behavior was different from pressing Escape e.g. in Vim hold down the
'i' key, and while the 'i' is getting printed now press 'escape'. The
default behavior is that the 'escape' key is honored and Vim returns to
'normal' mode.  However with the Hammerspoon implementation, the
'escape' was being interpreted as a 'control' press since the 'i' key
was still being held down.  Using the Karabiner implementation, things
work like in the default case.

Further the main motivating factor for this change was the very slow
reactive speed to using Vim remotely over SSH.