Fixes missing modifiers in newKeyEvent call
This issue ended up causing lots of pain, since it did not manifest
itself as a script error, nor did it impact the basic use case.  However
it did cause issues when modifier keys were being pressed or set by
other scripts.  So this is an important thing to look out for going forward.