Adds p10k.zsh
The p10k.zsh file is the generated config file from running `p10k
configure`. This file is from running it on my personal Macbook.
Probably this config is machine specific, but I am adding it to the
dotfiles repo so that there is an already available config for the
Docker image i.e. it removes the need for the manual run on the Docker
image for testing the ZSH config files. On machines that need a
different config, the `configure` command can be run again (without the
need to checkin the changes).