Fix ssh agent usage
Note online docs mention using '--mount=type=ssh' but don't mention that
this defaults to the root user.  After lots of searching around for why
I was still running into a permissions issue, I came across this info,
and that one can use the 'uid' parameter for a non-root user.