Adds non system OpenSSL to path
Needed to use non-sys-openssl on macOS for setting up email in Emacs. I
am not entirely sure why the non-system needs to be used, but all the
guides seem to mention it.  But I did not end up putting the certs in
the typical openssl path location and just put them in my home
directory. So maybe we could make do without this, but its better to
have this commit message explaining things than just silence and
potential later issues.