Update aliases
The alias `e` simply points to `$EDITOR` which is environment dependent
e.g. in a terminal `e` will open `emacsclient` in a terminal window (if
an Emacs daemon is not running it will fallback to NeoVim), but if `e`
is run from a terminal inside of Emacs, then it will open `emacsclient`
resulting in the file/directory opening in that instance (in particular
useful when the Emacs instance is a GUI instance).