Add SSH command back
Looks like the PATH variable in a terminal is not the one getting using
by `git.exe`. We need to go back to using an explicit SSH command
setting, and for now we will install Win32 OpenSSH so that the path is
available on all Windows machines that we use.