Using direnv instead of autoenv
I initially started with direnv and then moved to using autoenv since
autoenv supported defining functions and aliases, whereas direnv does
not given that it uses a sub-shell to execute .envrc files.

I am moving back to direnv as the mechanism it uses is really the same
method that has to be used for using Emacs in GUI mode (on macOS) since
Emacs is not a shell.  I am using the emacs-direnv package which enables
me to use Emacs in GUI mode without having to write my Elisp functions
to write the env to a file and load it into Emacs.

How do we handle the issues of not being able to export functions and
aliases?
    - Functions: we can use executable scripts, with the path of the
    scripts being added to PATH in the .envrc. Note: this is also how
    the "functions" are exposed to Emacs or any non-shell process.
    - Aliases: we kinda loose these... but maybe we are better of not
    using aliases as they are not good for sharing commands with others,
    and we can rather use completion tools for still quickly invoking commands.