Remove rg alias
This alias was originally added to produce nice color output, but I
feel that it is better to have it work as normal so that any demos
done with the `rg` command can have consisten results.  Also it makes
sense for terminal applications to not add color code sequences when
piping output to another program, so we should not be trying to
workaround this.